feat: cost-model + fee field compat for connector-go v0.4.0
Minimal API-compat changes on top of the pv3 commit the executor already
pins (d169a805), so callers built against upstream apollo v1.7.0+ (e.g.
cardano-connector-go v0.4.0) compile and run against this fork without
adopting newer pv3 changes or a full upstream rebase:

- ProtocolParameters.MinFeeConstant / MinFeeCoefficient: int -> int64
- ProtocolParameters.CostModels (map[string][]int64) renamed to
  CostModelsRaw, keeping json tag "cost_models" (connector-go v0.4.0
  writes the raw cost-model map into CostModelsRaw)

Internal consumers updated for the int64 fee fields (Base.Fee,
ToBaseParams, Utils.Fee, Maestro/Ogmios contexts) and the rename (Ogmios).
script_data_hash still uses the hardcoded PLUTUSV1/V2/V3 cost-model
constants exactly as before, so tx-building behaviour is unchanged.

+++ txBuilding/Backend/MaestroChainContext/MaestroChainContext.go
@@ -134,8 +134,8 @@ func (mcc *MaestroChainContext) LatestEpochParams() (Base.ProtocolParameters, er
 		return protocolParams, err
 	}
 	// Map ALL the fields
-	protocolParams.MinFeeConstant = int(ppFromApi.Data.MinFeeConstant.LovelaceAmount.Lovelace)
-	protocolParams.MinFeeCoefficient = int(ppFromApi.Data.MinFeeCoefficient)
+	protocolParams.MinFeeConstant = int64(ppFromApi.Data.MinFeeConstant.LovelaceAmount.Lovelace)
+	protocolParams.MinFeeCoefficient = int64(ppFromApi.Data.MinFeeCoefficient)
 	protocolParams.MaxTxSize = int(ppFromApi.Data.MaxTransactionSize.Bytes)
 	protocolParams.MaxBlockSize = int(ppFromApi.Data.MaxBlockBodySize.Bytes)
 	protocolParams.MaxBlockHeaderSize = int(ppFromApi.Data.MaxBlockHeaderSize.Bytes)
